metaconfoundr: Visualize Confounder Control in Meta-Analyses

Visualize confounder control in meta-analysis. 'metaconfoundr' is an approach to evaluating bias in studies used in meta-analyses based on the causal inference framework. Study groups create a causal diagram displaying their assumptions about the scientific question. From this, they develop a list of important confounders. Then, they evaluate whether studies controlled for these variables well. 'metaconfoundr' is a toolkit to facilitate this process and visualize the results as heatmaps, traffic light plots, and more.

See also

Author

Maintainer: Malcolm Barrett malcolmbarrett@gmail.com (ORCID)

Authors:

Julie M. Petersen juliemo@bu.edu (ORCID)

Ludovic Trinquart ludovic@bu.edu (ORCID)